feat(security): security to control is user logged, admin,visitor

+++ src/controller/BackController.php
@@ -6,54 +6,86 @@ use App\config\Parameter;
 
 class BackController extends Controller
 {
+    //Function to return if user is currently logged
+    private function checkLoggedIn()
+    {
+        if (!$this->session->get('pseudo')) {
+            $this->session->set('need_login', 'Vous devez vous connecter pour accéder à cette page');
+            header('Location: ../public/index.php?route=login');
+        } else {
+            return true;
+        }
+    }
+
+    //Function to return if a user is an admin
+    private function checkAdmin()
+    {
+        $this->checkLoggedIn();
+        if (!($this->session->get('role') === 'admin')) {
+            $this->session->set('not_admin', 'Vous n\'avez pas le droit d\'accéder à cette page');
+            header('Location: ../public/index.php?route=profile');
+        } else {
+            return true;
+        }
+    }
     //Function to return administration profile
     public function administration()
     {
-        $users = $this->userDAO->getUsers();
+        if ($this->checkAdmin()) {
+            $users = $this->userDAO->getUsers();
 
-        return $this->view->render('administration', [
-            'users' => $users
-        ]);
+            return $this->view->render('administration', [
+                'users' => $users
+            ]);
+        }
     }
 
     //Function to return profile view
     public function profile()
     {
-        return $this->view->render('profile');
+        if ($this->checkLoggedIn()) {
+            return $this->view->render('profile');
+        }
     }
 
     //Function to update user password
     public function updatePassword(Parameter $post)
     {
-        if ($post->get('submit')) {
+        if ($this->checkLoggedIn()) {
+            if ($post->get('submit')) {
 
-            $errors = $this->validation->validate($post, 'UpdatePassword');
+                $errors = $this->validation->validate($post, 'UpdatePassword');
 
-            if (!$errors) {
+                if (!$errors) {
 
-                $this->userDAO->updatePassword($post, $this->session->get('pseudo'));
-                $this->session->set('update_password', 'Le mot de passe a été mis à jour');
-                header('Location: ../public/index.php?route=profile');
+                    $this->userDAO->updatePassword($post, $this->session->get('pseudo'));
+                    $this->session->set('update_password', 'Le mot de passe a été mis à jour');
+                    header('Location: ../public/index.php?route=profile');
+                }
+                return  $this->view->render('update_password', [
+                    'post' => $post,
+                    'errors' => $errors
+                ]);
             }
-            return  $this->view->render('update_password', [
-                'post' => $post,
-                'errors' => $errors
-            ]);
+            return $this->view->render('update_password');
         }
-        return $this->view->render('update_password');
     }
 
     //Function to logout
     public function logout()
     {
-        $this->logoutOrDelete('logout');
+        if ($this->checkLoggedIn()) {
+            $this->logoutOrDelete('logout');
+        }
     }
 
     //Function to delete current user
     public function deleteAccount()
     {
-        $this->userDAO->deleteAccount($this->session->get('pseudo'));
-        $this->logoutOrDelete('delete_account');
+        if ($this->checkLoggedIn()) {
+            $this->userDAO->deleteAccount($this->session->get('pseudo'));
+            $this->logoutOrDelete('delete_account');
+        }
     }
 
     //Function to manage logout or delete 
@@ -72,8 +104,10 @@ class BackController extends Controller
     //Function to delete user
     public function deleteUser($userId)
     {
-        $this->userDAO->deleteUser($userId);
-        $this->session->set('delete_user', 'L\'utilisateur a bien été supprimé');
-        header('Location: ../public/index.php?route=administration');
+        if ($this->checkAdmin()) {
+            $this->userDAO->deleteUser($userId);
+            $this->session->set('delete_user', 'L\'utilisateur a bien été supprimé');
+            header('Location: ../public/index.php?route=administration');
+        }
     }
 }
